chat-server/llama_chat/requirements.txt
async-timeout==4.0.2
bidict==0.22.1
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
PyJWT==2.6.0
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
redis==4.5.2
six==1.16.0
SQLAlchemy==2.0.7
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
